mox-cu/03_boa-favorites_cu/pyproject.toml
[project]
name = "03-boa-favorites-cu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.1.0",
    "python-dotenv>=1.0.1",
    "titanoboa>=0.2.5",
]
